add distro check to base setup

+++ base_dependencies.sh
@@ -1,10 +1,19 @@
 #! /bin/bash
 
-echo "update packages"
-sudo apt update && sudo apt upgrade
+if [ -f /etc/debian_version ]; then
+    echo "update packages"
+    sudo apt update && sudo apt upgrade
 
-echo "install virtualenv"
-sudo apt install -y python3-virtualenv python3
+    echo "install virtualenv"
+    sudo apt install -y python3-virtualenv python3
+elif [ -f /etc/manjaro-release ]; then
+    echo "detected manjaro"
+    echo "update packages"
+    sudo pacman -Syu
+
+    echo "install virtualenv"
+    sudo pacman -S --noconfirm python-virtualenv python3
+fi
 
 echo "setup virtualenv"
 virtualenv env
